Sweeping changes

+++ src/tf_mmciad/utils/custom_loss.py
@@ -297,7 +297,7 @@ def categorical_focal_loss(gamma=2.0, alpha=0.25):
         y_pred = K.clip(y_pred, epsilon, 1.0 - epsilon)
 
         # Calculate Cross Entropy
-        cross_entropy = -y_true * K.log(y_pred)
+        cross_entropy = -tf.cast(y_true, dtype=tf.float32) * K.log(y_pred)
 
         # Calculate Focal Loss
         loss = alpha * K.pow(1 - y_pred, gamma) * cross_entropy
